move non testing logic to reference agent
